wip: porting to newly defined structure
