+++ src/common/entropy-base.ts
@@ -1,7 +1,7 @@
 import Entropy from "@entropyxyz/sdk";
 import { EntropyLogger } from "./logger";
 
-export abstract class BaseCommand {
+export abstract class EntropyBase {
   protected logger: EntropyLogger
   protected entropy: Entropy
